requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
bcrypt==4.0.1
SQLAlchemy==2.0.21
python-dotenv==1.0.0
Werkzeug==2.3.7 